refactor(#2453 Stage 1a): slim owner_services to production-only; move guards to app::tests
Addresses r0 review (test-bloat / KISS + over-broad claim):
- owner_services.rs is now PRODUCTION-ONLY (~47 lines): the ~150-line copied
  comment/string masker and the test module are removed.
- The three structural wiring guards move into app::tests, reusing the module's
  existing, proven `strip_comments_and_blank_strings` masker (no parser
  duplicated, no production visibility widened): owner_services_called_by_both_hosts,
  owner_services_spawns_absent_from_hosts_present_at_wiring_site,
  owner_services_calls_inside_attached_mode_guard. A small inline brace-range walk
  in the last one; a `mod tests {`-based prod-region cut (robust to daemon's
  item-level `#[cfg(test)]` and to an attribute before the module).
- The G2 claim is narrowed from "only inside owner_services" (which scans just the
  two hosts) to "absent from both host bodies AND present at the shared wiring
  site" — scope is the two run hosts + the wiring module, not global uniqueness.

Production extraction is unchanged (two helpers / five calls / exact positions /
exclusions). Guards stay STATIC source-scans — no helper called, no thread spawned;
runtime-host coverage intentionally ABSENT for this pure refactor.


Agend-Agent: archfix-claude-opus
Agend-Task: t-20260711201904711328-55009-3
Agend-Branch: refactor/2453-shared-owner-services
Agend-Issued-At: 2026-07-11T20:19:17.295465+00:00
